remove comments disable manage ban button for site mods when user is blanket banned disable all sites option when user is blanket banned
